+++ Plutarch/Convert.hs
@@ -16,8 +16,8 @@ module Plutarch.Convert (
 
 import GHC.Generics (Generic)
 import Plutarch.Builtin.Bool (PBool (PFalse, PTrue))
+import Plutarch.Builtin.Integer (PInteger)
 import Plutarch.ByteString (PByteString)
-import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Eq (PEq)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
 import Plutarch.Internal.Ord (POrd, PPartialOrd)