+++ plutarch-testlib/Plutarch/Test/Laws.hs
@@ -18,8 +18,10 @@ module Plutarch.Test.Laws (
 import Plutarch.Builtin (pforgetData)
 import Plutarch.Enum (PCountable (psuccessor, psuccessorN), PEnumerable (ppredecessor, ppredecessorN))
 import Plutarch.Internal.Lift (PLiftable (fromPlutarch, fromPlutarchRepr, toPlutarch, toPlutarchRepr))
+import Plutarch.Internal.Numeric (
+  PNum (pabs, pnegate, psignum, (#*), (#+), (#-)),
+ )
 import Plutarch.LedgerApi.V1 qualified as V1
-import Plutarch.Num (PNum (pabs, pnegate, psignum, (#*), (#+), (#-)))
 import Plutarch.Positive (PPositive, Positive)
 import Plutarch.Prelude
 import Plutarch.Test.QuickCheck (checkHaskellEquivalent, checkHaskellEquivalent2)